login conseguido!!

+++ src/main/java/com/example/demo/Repository/entity/PagoEntity.java
@@ -26,7 +26,7 @@ public class PagoEntity {
 
 
     private Long id;
-    private int posicionamiento;
+    private float posicionamiento;
     private Long idsolicitud;
     @Override
     public boolean equals(Object obj) {
